Change Sections table Number data type

+++ ELearning.Domain/Entities/Section.cs
@@ -11,7 +11,7 @@ namespace ELearning.Domain.Entities
         }
 
         public int SectionId { get; set; }
-        public int Number { get; set; }
+        public short Number { get; set; }
 
         public int GroupId { get; set; }
         public Group Group { get; set; }
